requirements.txt
# =========================
# FastAPI Core
# =========================
fastapi
uvicorn[standard]

# =========================
# Database & ORM
# =========================
sqlmodel
sqlalchemy
psycopg2-binary

# =========================
# Validation & Config
# =========================
pydantic
pydantic-settings
python-dotenv
email-validator

# =========================
# Authentication & Security
# =========================
python-jose[cryptography]
passlib[bcrypt]

# =========================
# Agent Orchestration (Future)
# =========================
langgraph
langchain
langchain-openai

# =========================
# HTTP & Utilities
# =========================
httpx
python-multipart

# =========================
# Migrations
# =========================
alembic

# =========================
# PDF Generation (Sanction Letter)
# =========================
pypdf
reportlab

# =========================
# Logging & Monitoring
# =========================
loguru

google-generativeai
